requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
packaging==24.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
redis==5.0.7
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
